README.md
Ever wanted to download all the reviews for a Steam game? The API given by Steam is great, but for larger games... I found that it misses a majority of the reviews. Other scrapers run into performance and memory issues when it comes to large games with over 100,000 reviews. Is there a solution? This is an alright bet!